requirements.txt
asgiref==3.8.1
boto3==1.36.1
botocore==1.36.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
djangorestframework==3.15.2
idna==3.10
jmespath==1.0.1
pyasn1==0.6.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
s3transfer==0.11.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
